refactor(ws): move WsMessageHandler to msgHandler folder
